requirements.txt
Cython>=0.29.21
numba
librosa
numpy
scipy
torch
unidecode
pyopenjtalk
jamo
pypinyin
pypinyin_dict
python-dotenv
jieba
protobuf
cn2an
inflect
eng_to_ipa
ko_pron
indic_transliteration
num_thai
opencc
uvicorn
fastapi
pydantic
loguru
soundfile
rtoml
graiax-silkcoder[libsndfile]
psutil
fasttext-wheel